Started DOM manipulation
Added css and html to begin the grid for tic tac toe. A problem I am now
having is something that has came up before. My array of players is
connected to a prompt which takes user input for the player name. I
want the intro to come up when I want, yet the script won't run unless
a value has been declared for the array player name.

How to wait for a variable to be declared until later ?