Added DOM that matches up with console game
Things of note:

Type error usually means that I am trying to do a method
to an element that requires something different. Like trying to do an
array method on a object.

It may be better to pass an argument to a function in its full form,
that way you have the option to access parts of it through dot notation.

Still don't understand when to use IIFE or factory.

Still having trouble managing the scope of things.